Add embedded prom server to container
Adds prom directly to the Dockerfile for my app.

This enables container to expose two ports, one for the app and one for the prom server

But managing the prom dep is very hacky right now and unclear how that would be made better.

Service disco + many open ports feels nicer ATM